+++ system/src/Grav/Common/Processors/DebuggerInitProcessor.php
@@ -8,13 +8,13 @@
 
 namespace Grav\Common\Processors;
 
-class DebuggerInitProcessor extends ProcessorBase implements ProcessorInterface {
-
+class DebuggerInitProcessor extends ProcessorBase implements ProcessorInterface
+{
     public $id = '_debugger';
     public $title = 'Init Debugger';
 
-    public function process() {
-      	$this->container['debugger']->init();
+    public function process()
+    {
+        $this->container['debugger']->init();
     }
-
 }