+++ system/src/Grav/Common/Processors/TwigProcessor.php
@@ -8,13 +8,14 @@
 
 namespace Grav\Common\Processors;
 
-class TwigProcessor extends ProcessorBase implements ProcessorInterface {
-
+class TwigProcessor extends ProcessorBase implements ProcessorInterface
+{
     public $id = 'twig';
     public $title = 'Twig';
 
-    public function process() {
-      	$this->container['twig']->init();
+    public function process()
+    {
+        $this->container['twig']->init();
     }
 
 }