+++ system/src/Grav/Common/Service/AssetsServiceProvider.php
@@ -14,7 +14,8 @@ use Grav\Common\Assets;
 
 class AssetsServiceProvider implements ServiceProviderInterface
 {
-    public function register(Container $container) {
+    public function register(Container $container)
+    {
         $container['assets'] = new Assets();
     }
 }